requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

pypdf
beautifulsoup4
lxml
chromadb
duckduckgo-search
ddgs
langchain-experimental
pinecone
langchain-pinecone
sentence-transformers
langchain-chroma